Move PaymentDeepLinkHandler to types file
