+++ packages/mobile/src/app/saga.ts
@@ -45,9 +45,9 @@ import { FETCH_TIMEOUT_DURATION } from 'src/config'
 import { SuperchargeTokenConfig } from 'src/consumerIncentives/types'
 import { handleDappkitDeepLink } from 'src/dappkit/dappkit'
 import { appVersionDeprecationChannel, fetchRemoteConfigValues } from 'src/firebase/firebase'
-import { PaymentDeepLinkHandler } from 'src/firebase/remoteConfigValuesDefaults'
 import { receiveAttestationMessage } from 'src/identity/actions'
 import { CodeInputType } from 'src/identity/verification'
+import { PaymentDeepLinkHandler } from 'src/merchantPayment/types'
 import { navigate } from 'src/navigator/NavigationService'
 import { Screens } from 'src/navigator/Screens'
 import { StackParamList } from 'src/navigator/types'
