+++ packages/mobile/src/firebase/remoteConfigValuesDefaults.e2e.ts
@@ -1,10 +1,6 @@
 import { RemoteConfigValues } from 'src/app/saga'
 import { SuperchargeButtonType } from 'src/app/types'
-
-export enum PaymentDeepLinkHandler {
-  Disabled = '',
-  Merchant = 'merchant',
-}
+import { PaymentDeepLinkHandler } from 'src/merchantPayment/types'
 
 export const REMOTE_CONFIG_VALUES_DEFAULTS: Omit<
   RemoteConfigValues,
