+++ packages/mobile/src/firebase/remoteConfigValuesDefaults.ts
@@ -1,11 +1,7 @@
 import { RemoteConfigValues } from 'src/app/saga'
 import { SuperchargeButtonType } from 'src/app/types'
 import { DEFAULT_SENTRY_NETWORK_ERRORS, DEFAULT_SENTRY_TRACES_SAMPLE_RATE } from 'src/config'
-
-export enum PaymentDeepLinkHandler {
-  Disabled = '',
-  Merchant = 'merchant',
-}
+import { PaymentDeepLinkHandler } from 'src/merchantPayment/types'
 
 export const REMOTE_CONFIG_VALUES_DEFAULTS: Omit<
   RemoteConfigValues,
