+++ packages/mobile/src/merchantPayment/types.ts
@@ -1,0 +1,4 @@
+export enum PaymentDeepLinkHandler {
+  Disabled = '',
+  Merchant = 'merchant',
+}
