+++ packages/mobile/src/qrcode/utils.ts
@@ -7,10 +7,10 @@ import { SendOrigin, WalletConnectPairingOrigin } from 'src/analytics/types'
 import ValoraAnalytics from 'src/analytics/ValoraAnalytics'
 import { ErrorMessages } from 'src/app/ErrorMessages'
 import { numberVerifiedSelector, paymentDeepLinkHandlerSelector } from 'src/app/selectors'
-import { PaymentDeepLinkHandler } from 'src/firebase/remoteConfigValuesDefaults'
 import i18n from 'src/i18n'
 import { validateRecipientAddressSuccess } from 'src/identity/actions'
 import { E164NumberToAddressType } from 'src/identity/reducer'
+import { PaymentDeepLinkHandler } from 'src/merchantPayment/types'
 import { navigate } from 'src/navigator/NavigationService'
 import { Screens } from 'src/navigator/Screens'
 import { UriData, uriDataFromUrl } from 'src/qrcode/schema'
